rename from friendly to nullsafe
